# docs/openapi/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AstraWatch Unified API",
    "description": "Official OpenAPI 3.0 specification for AstraWatch autonomous telemetry, anomaly detection, incident orchestration, and K8s self-healing microservices.",
    "version": "1.0.0",
    "contact": {
      "name": "AstraWatch Engineering",
      "url": "https://astrawatch.io"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8080/v1",
      "description": "Collector Service (:8080)"
    },
    {
      "url": "http://localhost:8000/v1",
      "description": "Analyzer Service (:8000)"
    },
    {
      "url": "http://localhost:8082/api/v1",
      "description": "Orchestrator Service (:8082)"
    },
    {
      "url": "http://localhost:8081/healthz",
      "description": "K8s Operator Service (:8081 healthz)"
    },
    {
      "url": "http://localhost:8084",
      "description": "Realtime Gateway (:8084)"
    },
    {
      "url": "http://localhost:8085/api/v1",
      "description": "Payment Service (:8085)"
    }
  ],
...
  "paths": {
...
    "/api/v1/incidents": {
      "post": {
        "tags": [
          "Orchestrator: Incidents"
        ],
        "summary": "Create Incident",
        "description": "Creates a new incident with severity level, affected service, and description.",
        "operationId": "oi1",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "id": "inc-001",
                  "severity": "HIGH",
                  "state": "OPEN",
                  "createdAt": "2026-01-01T00:00:00Z"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Incident payload",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "serviceId": "svc-001",
                "anomalyId": "anom-001",
                "severity": "HIGH",
                "title": "Payment service latency spike",
                "description": "P99 latency exceeds 5s threshold"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Orchestrator: Incidents"
        ],
        "summary": "List Incidents",
        "description": "Lists all incidents, optionally filtered by service.",
        "operationId": "oi2",
        "parameters": [
          {
            "name": "serviceId",
            "in": "query",
            "required": false,
            "description": "Filter by service",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": [
                  {
                    "id": "inc-001",
                    "title": "Payment latency",
                    "severity": "HIGH",
                    "state": "OPEN"
                  }
                ]
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/healing/history": {
      "get": {
        "tags": [
          "Orchestrator: Auto-Healing"
        ],
        "summary": "Healing History",
        "description": "Returns historical record of all executed healing actions.",
        "operationId": "oh4",
        "parameters": [
          {
            "name": "serviceId",
            "in": "query",
            "required": false,
            "description": "Service filter",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "actions": [
                    {
                      "id": "heal-001",
                      "actionType": "RESTART_POD",
                      "status": "COMPLETED",
                      "riskScore": 25
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/notifications/channels": {
      "get": {
        "tags": [
          "Orchestrator: Notifications"
        ],
        "summary": "List Channels",
        "description": "Lists all notification channels (Slack, PagerDuty, Email, Webhook).",
        "operationId": "on1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "channels": [
                    {
                      "id": "ch-001",
                      "name": "#alerts",
                      "type": "SLACK",
                      "isEnabled": true
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Orchestrator: Notifications"
        ],
        "summary": "Create Channel",
        "description": "Creates a new notification destination channel.",
        "operationId": "on2",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "id": "ch-002",
                  "name": "#prod-alerts",
                  "type": "SLACK"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Channel config",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "name": "#prod-alerts",
                "type": "SLACK",
                "config": "{ \"webhookUrl\": \"https://hooks.slack.com/...\" }",
                "isEnabled": true
              }
            }
          }
        }
      }
    },
    "/api/v1/notifications/channels/:id/test": {
      "post": {
        "tags": [
          "Orchestrator: Notifications"
        ],
        "summary": "Test Channel",
        "description": "Dispatches a test notification to verify channel integration.",
        "operationId": "on3",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Channel ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "delivered": true,
                  "responseCode": 200
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/notifications/maintenance-windows": {
      "get": {
        "tags": [
          "Orchestrator: Notifications"
        ],
        "summary": "List Maintenance Windows",
        "description": "Lists scheduled maintenance windows for notification suppression.",
        "operationId": "on4",
        "parameters": [
          {
            "name": "active",
            "in": "query",
            "required": false,
            "description": "Filter to active windows only",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "windows": [
                    {
                      "id": "mw-001",
                      "reason": "DB upgrade",
                      "startedAt": "2026-01-01T02:00:00Z",
                      "endedAt": "2026-01-01T04:00:00Z"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/api/v1/notifications/preferences": {
      "get": {
        "tags": [
          "Orchestrator: Notifications"
        ],
        "summary": "Get Preferences",
        "description": "Retrieves personal notification preferences for a user.",
        "operationId": "on5",
        "parameters": [
          {
            "name": "userId",
            "in": "query",
            "required": true,
            "description": "User ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": [
                  {
                    "channelType": "SLACK",
                    "severityMin": "HIGH",
                    "isEnabled": true,
                    "quietHoursStart": "22:00",
                    "quietHoursEnd": "07:00"
                  }
                ]
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/api/v1/rules": {
      "get": {
        "tags": [
          "Operator: AutoHealingRules"
        ],
        "summary": "List Rules",
        "description": "Returns all AutoHealingRule custom resources from the Kubernetes cluster.",
        "operationId": "opr1",
        "parameters": [
          {
            "name": "namespace",
            "in": "query",
            "required": false,
            "description": "K8s namespace (default: all)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "rules": [
                    {
                      "name": "restart-on-oom",
                      "namespace": "production",
                      "spec": {
                        "targetDeployment": "payment-service",
                        "trigger": "OOMKilled",
                        "action": "RESTART"
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "Operator: AutoHealingRules"
        ],
        "summary": "Create Rule",
        "description": "Applies a new AutoHealingRule CRD to the Kubernetes cluster.",
        "operationId": "opr2",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                },
                "example": {
                  "name": "restart-on-oom",
                  "status": "Created"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "description": "AutoHealingRule spec",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "example": {
                "name": "restart-on-oom",
                "namespace": "production",
                "spec": {
                  "targetDeployment": "payment-service",
                  "trigger": "OOMKilled",
                  "action": "RESTART_POD",
                  "blastRadius": 1,
                  "cooldownSeconds": 300
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter your Bearer JWT token issued by POST /api/v1/auth/login"
      }
    },
...
  }
}